util function and unit test
