+++ tests/unit/test_ml_utils.py
@@ -19,6 +19,34 @@ from sklearn.linear_model import LogisticRegression, SGDClassifier
 from dask_sql.physical.rel.custom.wrappers import Incremental, ParallelPostFit
 
 
+@pytest.mark.parametrize("gpu", [False, pytest.param(True, marks=pytest.mark.gpu)])
+def test_ml_class_mappings(gpu):
+    from dask_sql.physical.utils.ml_classes import get_cpu_classes, get_gpu_classes
+    from dask_sql.utils import import_class
+
+    try:
+        import lightgbm
+        import xgboost
+    except KeyError:
+        lightgbm = None
+        xgboost = None
+
+    if gpu:
+        classes_dict = get_gpu_classes()
+    else:
+        # Imports needed to use sklearn.experimental classes
+        from sklearn.experimental import enable_halving_search_cv  # noqa: F401
+        from sklearn.experimental import enable_iterative_imputer  # noqa: F401
+
+        classes_dict = get_cpu_classes()
+
+    for key in classes_dict:
+        if not ("XGB" in key and xgboost is None) and not (
+            "LGBM" in key and lightgbm is None
+        ):
+            import_class(classes_dict[key])
+
+
 def _check_axis_partitioning(chunks, n_features):
     c = chunks[1][0]
     if c != n_features:
